♻️ refactor!: update to typescript, relicense to MIT, do the uh thingy idk
