+++ .github/workflows/build.yml
@@ -1,0 +1,22 @@
+name: Lint commits
+on: [push, pull_request]
+
+permissions:
+  checks: write
+
+jobs:
+  lint:
+    name: Lint commits
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: pnpm/action-setup@v4
+      - uses: actions/setup-node@v4
+        with:
+          node-version: 22
+          cache: pnpm
+
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+      - name: Try to build
+        run: pnpm build
