+++ README.md
@@ -1,0 +1,3 @@
+# `@moonlight-mod/esbuild-config`
+
+A simple and effective `esbuild` configuration for moonlight extensions and builds.
