+++ package.json
@@ -1,10 +1,39 @@
 {
   "name": "@moonlight-mod/esbuild-config",
   "version": "1.0.0",
-  "main": "index.js",
   "type": "module",
+  "private": false,
+  "packageManager": "pnpm@10.7.1",
+  "description": "Yet another Discord mod's shared esbuild config",
+  "license": "LGPL-3.0-or-later",
+  "homepage": "https://moonlight-mod.github.io/",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/moonlight-mod/esbuild-config.git"
+  },
+  "bugs": {
+    "url": "https://github.com/moonlight-mod/esbuild-config/issues"
+  },
+  "engineStrict": true,
+  "engines": {
+    "node": ">=22",
+    "pnpm": ">=10",
+    "npm": "pnpm",
+    "yarn": "pnpm"
+  },
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "build": "tsup",
+    "dev": "tsup --watch",
+    "updates": "pnpm taze -r"
+  },
   "devDependencies": {
-    "prettier": "^3.4.2"
+    "@types/node": "^22.14.0",
+    "prettier": "^3.4.2",
+    "esbuild": "^0.25.2",
+    "typescript": "^5.8.3",
+    "tsup": "^8.4.0"
   },
   "peerDependencies": {
     "esbuild": ">= 0.19"
