--- plugins/deduplicatedLogging.js
@@ -1,31 +1,0 @@
-import * as esbuild from "esbuild";
-
-let lastMessages = new Set();
-
-/** @type {import("esbuild").Plugin} */
-const deduplicatedLogging = {
-  name: "deduplicated-logging",
-  setup(build) {
-    build.onStart(() => {
-      lastMessages.clear();
-    });
-
-    build.onEnd(async (result) => {
-      const formatted = await Promise.all([
-        esbuild.formatMessages(result.warnings, {
-          kind: "warning",
-          color: true
-        }),
-        esbuild.formatMessages(result.errors, { kind: "error", color: true })
-      ]).then((a) => a.flat());
-
-      for (const message of formatted) {
-        if (lastMessages.has(message)) continue;
-        lastMessages.add(message);
-        console.log(message.trim());
-      }
-    });
-  }
-};
-
-export default deduplicatedLogging;
