+++ tsconfig.json
@@ -1,0 +1,35 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig.json",
+    "compilerOptions": {
+        "target": "ESNext",
+        "jsx": "preserve",
+        "lib": [
+            "ESNext",
+            "ESNext.Disposable"
+        ],
+        "module": "ESNext",
+        "moduleResolution": "Bundler",
+        "resolveJsonModule": true,
+        "allowArbitraryExtensions": false,
+        "noEmit": true,
+        "allowImportingTsExtensions": true,
+        "allowJs": true,
+        "strict": true,
+        "strictNullChecks": true,
+        "noFallthroughCasesInSwitch": true,
+        "noImplicitReturns": true,
+        "noErrorTruncation": true,
+        "declaration": true,
+        "declarationMap": true,
+        "outDir": "dist",
+        "sourceMap": true,
+        "stripInternal": true,
+        "esModuleInterop": true,
+        "verbatimModuleSyntax": true
+    },
+    "exclude": [
+        "**/node_modules/**",
+        "**/dist/**",
+        "**/build/**"
+    ]
+}
